Split `exit_code.rs` into submodules
