+++ CHANGELOG.adoc
@@ -14,6 +14,12 @@ All notable changes to this project will be documented in this file.
 The format is based on https://keepachangelog.com/[Keep a Changelog], and this
 project adheres to https://semver.org/[Semantic Versioning].
 
+== {compare-url}/v0.7.13\...HEAD[Unreleased]
+
+=== Changed
+
+* Split `exit_code.rs` into submodules ({pull-request-url}/87[#87])
+
 == {compare-url}/v0.7.12\...v0.7.13[0.7.13] - 2024-04-16
 
 === Added
